added branch 'longer'

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        SalesManager sm = new SalesManager(new int[]{3, 6, 9});
+        SalesManager sm = new SalesManager(new long[]{3, 6, 9});
         System.out.println(sm.max());
     }
 }
